+++ options/parse_configfile.h
@@ -18,7 +18,7 @@
 #ifndef MPLAYER_PARSER_CFG_H
 #define MPLAYER_PARSER_CFG_H
 
-#include "m_config.h"
+#include "m_config_frontend.h"
 
 int m_config_parse_config_file(m_config_t* config, const char *conffile,
                                char *initial_section, int flags);
